+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version history
 
+- 0.3.0.2 – (2023/05/24) COPY stmt and SYMBOLs file (var address) bug fix;
+- 0.3.0.1 – (2023/05/23) COPY screen data in tiled mode bug fix;
 - 0.3.0.0 – (2023/02/17) TURBO function bug fix, integer math optimizations;
 - 0.2.9.9 – (2023/01/20) SET FONT new command;
 - 0.2.9.8 – (2022/12/21) MegaROM mode support for CMD, RESOURCES and Arkos Tracker player;